Update Readme
This mod aims to replace Ambience and Actual Music as one that works without too many issues and no crashes to speak of. All handling of the music is done entirely through the config, which means the tracks you add there become actual registered sounds in game. The tracks are even put on music discs for you! Refer to the Curseforge description to see how to use the mod